backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
numpy
scikit-learn
starlette
jinja2
